I think this is right
CMC9 is Z and Y direction periodicity.
VSPT is only Y direction
